apps/ai-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
numpy>=1.24.3
opencv-python>=4.8.1.78
librosa>=0.10.0
torch>=2.1.1
torchaudio>=2.1.1
transformers>=4.35.0
sounddevice>=0.4.6
scipy>=1.11.4
deepface>=0.0.75
openai-whisper>=20231117
pydantic-settings>=2.1.0
httpx>=0.25.2
requests>=2.31.0
